fix Kinesis It
